requirements.txt
alembic==1.8.0
aniso8601==9.0.1
attrs==21.4.0
click==8.1.3
Flask==2.1.2
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
Mako==1.2.0
MarkupSafe==2.1.1
PyMySQL==1.0.2
pyrsistent==0.18.1
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.38
Werkzeug==2.1.2
zipp==3.8.0
